refactor: Explode Oracle
