+++ packages/prophet/Oracle/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default } from "./Oracle";
